+++ src/main/kotlin/com/cycode/plugin/Consts.kt
@@ -27,7 +27,7 @@ class Consts {
     companion object {
         val PLUGIN_PATH = PathManager.getPluginsPath() + "/cycode-intellij-platform-plugin"
         val DEFAULT_CLI_PATH = getDefaultCliPath()
-        const val REQUIRED_CLI_VERSION = "2.1.0"
+        const val REQUIRED_CLI_VERSION = "2.2.0"
 
         const val CYCODE_DOMAIN = "cycode.com"