+++ src/main/kotlin/com/cycode/plugin/cli/CliIgnoreType.kt
@@ -3,5 +3,6 @@ package com.cycode.plugin.cli
 enum class CliIgnoreType {
     VALUE,
     RULE,
-    PATH
+    PATH,
+    CVE,
 }