+++ src/main/kotlin/com/cycode/plugin/services/scanResultsFilters/ScanResultsFilterBase.kt
@@ -1,7 +1,7 @@
 package com.cycode.plugin.services.scanResultsFilters
 
 abstract class ScanResultsFilterBase<T>(val scanResults: T) {
-    fun exclude(byValue: String? = null, byPath: String? = null, byRuleId: String? = null) {
+    fun exclude(byValue: String? = null, byPath: String? = null, byRuleId: String? = null, byCve: String? = null) {
         if (byValue != null) {
             excludeByValue(byValue)
         }
@@ -11,6 +11,9 @@ abstract class ScanResultsFilterBase<T>(val scanResults: T) {
         if (byRuleId != null) {
             excludeByRuleId(byRuleId)
         }
+        if (byCve != null) {
+            excludeByCve(byCve)
+        }
     }
 
     abstract fun excludeByValue(value: String)
@@ -18,6 +21,8 @@ abstract class ScanResultsFilterBase<T>(val scanResults: T) {
     abstract fun excludeByPath(path: String)
 
     abstract fun excludeByRuleId(ruleId: String)
+
+    abstract fun excludeByCve(cve: String)
 
     abstract fun getFilteredScanResults(): T
 }